+++ backend/route/message-router.js
@@ -4,7 +4,7 @@ const debug = require('debug')('Backend-Portfolio:message-router.js');
 
 const Router = require('express').Router;
 const jsonParser = require('body-parser').json();
-const Promise = require('bluebird');
+// const Promise = require('bluebird');
 const createError = require('http-errors');
 
 const bearerAuth = require('../lib/bearer-auth-middleware.js');
@@ -15,10 +15,12 @@ const messageRouter = module.exports = Router();
 messageRouter.post('/api/message', bearerAuth, jsonParser, (req, res, next) => {
   debug('POST: /api/message');
 
-  if(!req.body || !req.body.authorId) return next(createError(400, 'missing id credentials'));
+  if(!req.body || (!req.body.photos && (!req.body.title && !req.body.text))) return next(createError(400, 'missing minimum content requirments'));
+  if(!req.user || !req.user._id) return next(createError(401, 'json web token failure, unauthorized'));
 
   new Message({
-    authorId: req.body.authorId,
+    // NOTE: this is probably a bad idea to have a user's id within each message
+    authorId: req.user._id,
     text: req.body.text || null,
     title: req.body.title || null,
     photos: req.body.photos || [],
@@ -37,41 +39,46 @@ messageRouter.get('/api/message/all', bearerAuth, (req, res, next) => {
     .catch(next);
 });
 
-messageRouter.get('/api/message/self/:id', bearerAuth, (req, res, next) => {
-  debug('GET: /api/message/self/:id');
+messageRouter.get('/api/message/self', bearerAuth, (req, res, next) => {
+  debug('GET: /api/message/self');
 
-  if(!req.params || !req.params.id) return next(createError(400, 'missing authorId field'));
+  if(!req.user || !req.user._id) return next(createError(401, 'json web token failure, unauthorized'));
 
-  Message.find({authorId: req.params.id})
-    .then((messages) => res.json(messages))
+  Message.find({authorId: req.user._id})
+    .then((messages) => {
+      if(!messages) res.status(204).send();
+      res.json(messages);
+    })
     .catch(next);
 });
 
+// NOTE: I think this one should stay as it is
 messageRouter.get('/api/message/:id', bearerAuth, (req, res, next) => {
   debug('GET: /api/message/:id');
 
   if(!req.params.id) return next(createError(400, 'missing user\'s request id'));
 
-  Message.findById(req.params.id)
+  Message.findById({_id: req.params.id})
     .then((message) => res.json(message))
     .catch(next);
 });
 
-// NOTE: should really add a catch to make sure you can't edit other people's photos
 messageRouter.put('/api/message/edit/:id', bearerAuth, jsonParser, (req, res, next) => {
   debug('PUT: /api/message/edit/:id');
 
-  if(!req.params.id) return next(createError(400, 'missing user\'s request id'));
+  if(!req.params || !req.params.id) return next(createError(400, 'missing user\'s request id'));
+  if(!req.user || !req.user._id) return next(createError(401, 'json web token failure, unauthorized'));
 
-  Message.findById(req.params.id)
+  Message.findById({_id: req.params.id})
     .then((message) => {
+      if(message.authorId !== req.user._id) return next(createError(401, 'you are not authorized to edit this post'));
       if(req.body.text){
         message.text = req.body.text;
       }
       if(req.body.title){
         message.title = req.body.title;
       }
-      // NOTE: this may need to be fixed later
+      // NOTE: this is broken
       if(req.body.photos){
         message.photos = req.body.photos;
       }
@@ -80,14 +87,23 @@ messageRouter.put('/api/message/edit/:id', bearerAuth, jsonParser, (req, res, ne
     .catch(next);
 });
 
-// NOTE: found another small issue, users can remove other people's comments, need to add an id checker I think
 // NOTE: I think I need to remove all comments associated with each message upon deletion
 messageRouter.delete('/api/message/remove/:id', bearerAuth, (req, res, next) => {
   debug('DELETE: /api/message/remove/:id');
 
   if(!req.params.id) return next(createError(400, 'missing user\'s request id'));
-
-  Message.findByIdAndRemove(req.params.id)
-    .then(() => res.status(204).send())
+  if(!req.user || !req.user._id) return next(createError(401, 'you are not authorized to remove this post'));
+  Message.findById({_id: req.params.id})
+    .then((message) => {
+      console.log(req.user._id)
+      console.log(typeof JSON.stringify(req.user._id))
+      console.log(message.authorId)
+      console.log(typeof message.authorId)
+      if(!message) return next(createError(404, 'no message was found with this id'));
+      if(message.authorId !== JSON.stringify(req.user._id)) return next(createError(401, 'you are not authorized to remove this post'));
+      Message.findByIdAndRemove(req.params.id)
+        .then(() => res.status(204).send())
+        .catch(next);
+    })
     .catch(next);
 });
